DoctorAPP/back-end/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-dotenv>=1.0.0
langchain>=0.3.0
langchain-openai>=0.3.0
sentence-transformers>=3.0.0
pymongo[srv]>=4.10.0
httpx>=0.28.0
pydantic>=2.10.0
xrpl-py>=4.0.0
aiosmtplib>=3.0.0
